+++ tests/pq/test_pgconn.py
@@ -1,22 +1,26 @@
 from __future__ import annotations
 
+import contextlib
 import os
 import sys
+import time
 import ctypes
 import logging
 import weakref
+from functools import partial
 from select import select
+from typing import Iterator
 
 import pytest
 
 import psycopg
 from psycopg import pq
-from psycopg.pq.abc import PGconn
+from psycopg.pq.abc import PGcancelConn, PGconn
 import psycopg.generators
 
 
 def wait(
-    conn: PGconn,
+    conn: PGconn | PGcancelConn,
     poll_method: str = "connect_poll",
     return_on: pq.PollingStatus = pq.PollingStatus.OK,
     timeout: int | None = None,
@@ -379,6 +383,87 @@ def test_set_single_row_mode(pgconn):
 
     pgconn.send_query(b"select 1")
     pgconn.set_single_row_mode()
+
+
+@contextlib.contextmanager
+def cancellable_query(pgconn: PGconn, monitor_conn: PGconn) -> Iterator[None]:
+    pgconn.send_query_params(b"SELECT pg_sleep($1)", [b"180"])
+    while True:
+        r = monitor_conn.exec_(
+            b"SELECT count(*) FROM pg_stat_activity"
+            b" WHERE query = 'SELECT pg_sleep($1)'"
+            b" AND state = 'active'"
+        )
+        assert r.status == pq.ExecStatus.TUPLES_OK
+        if r.get_value(0, 0) != b"0":
+            break
+
+        time.sleep(0.01)
+
+    yield None
+
+    res = pgconn.get_result()
+    assert res is not None
+    assert res.status == pq.ExecStatus.FATAL_ERROR
+    assert res.error_field(pq.DiagnosticField.SQLSTATE) == b"57014"
+    while pgconn.is_busy():
+        pgconn.consume_input()
+
+
+@pytest.mark.libpq(">= 17")
+def test_cancel_nonblocking(dsn):
+    # mimic test_cancel() from src/test/modules/libpq_pipeline/libpq_pipeline.c
+    def connect() -> PGconn:
+        conn = pq.PGconn.connect(dsn.encode())
+        if conn.status != pq.ConnStatus.OK:
+            pytest.fail(
+                f"bad connection: {conn.error_message.decode('utf8', 'replace')}"
+            )
+        return conn
+
+    conn, monitor_conn = connect(), connect()
+    conn.nonblocking = 1
+
+    # test PQcancel
+    with cancellable_query(conn, monitor_conn):
+        cancel = conn.get_cancel()
+        cancel.cancel()
+
+    # PGcancel object can be reused for the next query
+    with cancellable_query(conn, monitor_conn):
+        cancel.cancel()
+
+    del cancel
+
+    # test PQcancelBlocking
+    with cancellable_query(conn, monitor_conn):
+        cancel_conn = conn.cancel_conn()
+        assert cancel_conn.status == pq.ConnStatus.ALLOCATED
+        cancel_conn.blocking()
+        assert cancel_conn.status == pq.ConnStatus.OK  # type: ignore[comparison-overlap] # noqa: E501
+    cancel_conn.finish()
+    del cancel_conn
+
+    wait_cancel = partial(wait, poll_method="poll", timeout=3)
+
+    # test PQcancelCreate and then polling with PQcancelPoll
+    with cancellable_query(conn, monitor_conn):
+        cancel_conn = conn.cancel_conn()
+        assert cancel_conn.status == pq.ConnStatus.ALLOCATED
+        cancel_conn.start()
+        assert cancel_conn.status == pq.ConnStatus.STARTED
+        wait_cancel(cancel_conn)
+        assert cancel_conn.status == pq.ConnStatus.OK
+
+    # test PQcancelReset works on the cancel connection and it can be reused
+    # after
+    cancel_conn.reset()
+    with cancellable_query(conn, monitor_conn):
+        cancel_conn.start()
+        wait_cancel(cancel_conn)
+        assert cancel_conn.status == pq.ConnStatus.OK
+
+    cancel_conn.finish()
 
 
 def test_cancel(pgconn):